requirements.txt
streamlit
gspread
google-auth
firebase-admin
fpdf